agents/microsoft-framework/requirements.txt
# Microsoft Agent Framework Dependencies
# Latest preview packages for advanced agent capabilities

# Core Agent Framework
agent-framework-azure-ai>=0.1.0a1  # --pre flag required during preview

# Azure SDK Integration
azure-ai-projects>=1.0.0
azure-identity>=1.15.0
azure-core>=1.29.0

# OpenTelemetry Tracing (Required for observability)
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp>=1.21.0
opentelemetry-instrumentation>=0.42b0

# Structured Output & Validation
pydantic>=2.5.0
jsonschema>=4.20.0

# Configuration Management
pydantic-settings>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# HTTP Client for MCP Integration
httpx>=0.25.0
aiohttp>=3.9.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
mypy>=1.7.0

# Optional: Enhanced capabilities
# numpy>=1.24.0  # For embeddings/vector operations
# pandas>=2.0.0  # For data processing workflows